+++ backend/substrapp/compute_tasks/errors.py
@@ -1,7 +1,6 @@
 """Objects to manage errors occurring in a compute task."""
 
 import enum
-from io import BytesIO
 from typing import BinaryIO
 
 from orchestrator import failure_report_pb2
@@ -61,34 +60,6 @@ class _ComputeTaskError(RuntimeError):
     error_type: ComputeTaskErrorType
 
 
-class BuildRetryError(_ComputeTaskError, CeleryRetryError):
-    """An error occurred during the build of a container image.
-
-    Args:
-        logs (str): the container image build logs
-    """
-
-    error_type = ComputeTaskErrorType.BUILD_ERROR
-
-    def __init__(self, logs: str, *args, **kwargs):
-        self.logs = BytesIO(str.encode(logs))
-        super().__init__(logs, *args, **kwargs)
-
-
-class BuildError(_ComputeTaskError, CeleryNoRetryError):
-    """An error occurred during the build of a container image.
-
-    Args:
-        logs (str): the container image build logs
-    """
-
-    error_type = ComputeTaskErrorType.BUILD_ERROR
-
-    def __init__(self, logs: str, *args, **kwargs):
-        self.logs = BytesIO(str.encode(logs))
-        super().__init__(logs, *args, **kwargs)
-
-
 class ExecutionError(_ComputeTaskError, CeleryNoRetryError):
     """An error occurred during the execution of a command in a container image.
 
@@ -100,10 +71,10 @@ class ExecutionError(_ComputeTaskError, CeleryNoRetryError):
 
     def __init__(self, logs: BinaryIO, *args, **kwargs):
         self.logs = logs
-        super().__init__(*args, **kwargs)
+        super().__init__(logs, *args, **kwargs)
 
 
-def get_error_type(exc: Exception) -> failure_report_pb2.ErrorType:
+def get_error_type(exc: Exception) -> failure_report_pb2.ErrorType.ValueType:
     """From a given exception, return an error type safe to store and to advertise to the user.
 
     Args: