+++ backend/substrapp/models/__init__.py
@@ -1,8 +1,10 @@
-from .compute_task_failure_report import ComputeTaskFailureReport
+from .asset_failure_report import AssetFailureReport
+from .asset_failure_report import FailedAssetKind
 from .computeplan_worker_mapping import ComputePlanWorkerMapping
 from .datamanager import DataManager
 from .datasample import DataSample
 from .function import Function
+from .function import FunctionImage
 from .image_entrypoint import ImageEntrypoint
 from .model import Model
 from .worker_last_event import WorkerLastEvent
@@ -10,10 +12,12 @@ from .worker_last_event import WorkerLastEvent
 __all__ = [
     "DataSample",
     "DataManager",
+    "FailedAssetKind",
     "Function",
+    "FunctionImage",
     "Model",
     "ComputePlanWorkerMapping",
     "ImageEntrypoint",
-    "ComputeTaskFailureReport",
+    "AssetFailureReport",
     "WorkerLastEvent",
 ]