+++ backend/substrapp/tasks/tasks_compute_task.py
@@ -4,14 +4,12 @@ This file contains the main logic for executing a compute task:
 - Create execution context
 - Populate asset buffer
 - Loads assets from the asset buffer
-- Build container images
 - **Execute the compute task**
 - Save the models/results
 - Teardown the context
 
 We also handle the retry logic here.
 """
-
 from __future__ import annotations
 
 import datetime
@@ -19,30 +17,24 @@ import enum
 import errno
 import os
 from typing import Any
-from typing import Optional
 
 import celery.exceptions
 import structlog
-from billiard.einfo import ExceptionInfo
-from celery import Task
 from celery.result import AsyncResult
 from django.conf import settings
-from django.core import files
 from rest_framework import status
 
 import orchestrator
 from backend.celery import app
-from substrapp import models
-from substrapp import utils
 from substrapp.clients import organization as organization_client
 from substrapp.compute_tasks import compute_task as task_utils
 from substrapp.compute_tasks import errors as compute_task_errors
+from substrapp.compute_tasks import image_builder
 from substrapp.compute_tasks.asset_buffer import add_assets_to_taskdir
 from substrapp.compute_tasks.asset_buffer import add_task_assets_to_buffer
 from substrapp.compute_tasks.asset_buffer import clear_assets_buffer
 from substrapp.compute_tasks.asset_buffer import init_asset_buffer
 from substrapp.compute_tasks.chainkeys import prepare_chainkeys_dir
-from substrapp.compute_tasks.compute_pod import delete_compute_plan_pods
 from substrapp.compute_tasks.context import Context
 from substrapp.compute_tasks.datastore import Datastore
 from substrapp.compute_tasks.datastore import get_datastore
@@ -53,14 +45,15 @@ from substrapp.compute_tasks.directories import init_task_dirs
 from substrapp.compute_tasks.directories import restore_dir
 from substrapp.compute_tasks.directories import teardown_task_dirs
 from substrapp.compute_tasks.execute import execute_compute_task
-from substrapp.compute_tasks.image_builder import build_image_if_missing
 from substrapp.compute_tasks.lock import MAX_TASK_DURATION
 from substrapp.compute_tasks.lock import acquire_compute_plan_lock
 from substrapp.compute_tasks.outputs import OutputSaver
 from substrapp.exceptions import OrganizationHttpError
 from substrapp.lock_local import lock_resource
 from substrapp.orchestrator import get_orchestrator_client
+from substrapp.tasks.task import ComputeTask
 from substrapp.utils import Timer
+from substrapp.utils import get_owner
 from substrapp.utils import list_dir
 from substrapp.utils import retry
 from substrapp.utils.url import TASK_PROFILING_BASE_URL
@@ -76,67 +69,6 @@ class ComputeTaskSteps(enum.Enum):
     PREPARE_INPUTS = "prepare_inputs"
     TASK_EXECUTION = "task_execution"
     SAVE_OUTPUTS = "save_outputs"
-
-
-class ComputeTask(Task):
-    autoretry_for = settings.CELERY_TASK_AUTORETRY_FOR
-    max_retries = settings.CELERY_TASK_MAX_RETRIES
-    retry_backoff = settings.CELERY_TASK_RETRY_BACKOFF
-    retry_backoff_max = settings.CELERY_TASK_RETRY_BACKOFF_MAX
-    retry_jitter = settings.CELERY_TASK_RETRY_JITTER
-
-    @property
-    def attempt(self) -> int:
-        return self.request.retries + 1  # type: ignore
-
-    def on_success(self, retval: dict[str, Any], task_id: str, args: tuple, kwargs: dict[str, Any]) -> None:
-        from django.db import close_old_connections
-
-        close_old_connections()
-
-    def on_retry(self, exc: Exception, task_id: str, args: tuple, kwargs: dict[str, Any], einfo: ExceptionInfo) -> None:
-        _, task = self.split_args(args)
-        # delete compute pod to reset hardware ressources
-        delete_compute_plan_pods(task.compute_plan_key)
-        logger.info(
-            "Retrying task",
-            celery_task_id=task_id,
-            attempt=(self.attempt + 1),
-            max_attempts=(settings.CELERY_TASK_MAX_RETRIES + 1),
-        )
-
-    def on_failure(
-        self, exc: Exception, task_id: str, args: tuple, kwargs: dict[str, Any], einfo: ExceptionInfo
-    ) -> None:
-        from django.db import close_old_connections
-
-        close_old_connections()
-
-        channel_name, task = self.split_args(args)
-        compute_task_key = task.key
-
-        failure_report = _store_failure(exc, compute_task_key)
-        error_type = compute_task_errors.get_error_type(exc)
-
-        with get_orchestrator_client(channel_name) as client:
-            # On the backend, only execution errors lead to the creation of compute task failure report instances
-            # to store the execution logs.
-            if failure_report:
-                logs_address = {
-                    "checksum": failure_report.logs_checksum,
-                    "storage_address": failure_report.logs_address,
-                }
-            else:
-                logs_address = None
-
-            client.register_failure_report(
-                {"compute_task_key": compute_task_key, "error_type": error_type, "logs_address": logs_address}
-            )
-
-    def split_args(self, celery_args: tuple) -> tuple[str, orchestrator.ComputeTask]:
-        channel_name = celery_args[0]
-        task = orchestrator.ComputeTask.model_validate_json(celery_args[1])
-        return channel_name, task
 
 
 def queue_compute_task(channel_name: str, task: orchestrator.ComputeTask) -> None:
@@ -164,7 +96,10 @@ def queue_compute_task(channel_name: str, task: orchestrator.ComputeTask) -> Non
         worker_queue=worker_queue,
     )
 
-    compute_task.apply_async((channel_name, task, task.compute_plan_key), queue=worker_queue, task_id=task.key)
+    compute_task.apply_async(
+        (channel_name, task, task.compute_plan_key),
+        queue=worker_queue,
+    )
 
 
 @app.task(
@@ -262,7 +197,10 @@ def _run(
         # start build_image timer
         timer.start()
 
-        build_image_if_missing(datastore, ctx.function)
+        image_builder.wait_for_image_built(ctx.function, channel_name)
+
+        if get_owner() != ctx.function.owner:
+            image_builder.load_remote_function_image(ctx.function, channel_name)
 
         # stop build_image timer
         _create_task_profiling_step(channel_name, task.key, ComputeTaskSteps.BUILD_IMAGE, timer.stop())
@@ -339,23 +277,3 @@ def _run(
 def _prepare_chainkeys(compute_plan_dir: str, compute_plan_tag: str) -> None:
     chainkeys_dir = os.path.join(compute_plan_dir, CPDirName.Chainkeys)
     prepare_chainkeys_dir(chainkeys_dir, compute_plan_tag)  # does nothing if chainkeys already populated
-
-
-def _store_failure(exc: Exception, compute_task_key: str) -> Optional[models.ComputeTaskFailureReport]:
-    """If the provided exception is a `BuildError` or an `ExecutionError`, store its logs in the Django storage and
-    in the database. Otherwise, do nothing.
-
-    Returns:
-        An instance of `models.ComputeTaskFailureReport` storing the error logs or None if the provided exception is
-        neither a `BuildError` nor an `ExecutionError`.
-    """
-
-    if not isinstance(exc, (compute_task_errors.ExecutionError, compute_task_errors.BuildError)):
-        return None
-
-    file = files.File(exc.logs)
-    failure_report = models.ComputeTaskFailureReport(
-        compute_task_key=compute_task_key, logs_checksum=utils.get_hash(file)
-    )
-    failure_report.logs.save(name=compute_task_key, content=file, save=True)
-    return failure_report